# Tiff: Performance improvements for Fax4 decompression