--- a/tests/Images/Input/Tiff/CCITTGroup4.tiff
+++ b/tests/Images/Input/Tiff/CCITTGroup4.tiff
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:120ad0814f207c45d968b05f7435034ecfee8ac1a0958cd984a070dad31f66f3
+size 11082