### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR brings performance improvements for Tiff Fax4 decompression.

Related to #2132 

Benchmarks:

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19043.1706 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.100-preview.4.22252.9
  [Host]     : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT
  Job-NMOCEA : .NET 6.0.5 (6.0.522.21309), X64 RyuJIT

Runtime=.NET 6.0  Arguments=/p:DebugType=portable  InvocationCount=1
IterationCount=3  LaunchCount=1  UnrollFactor=1
WarmupCount=3
```

main 

```
|                Method |                                      TestImage |         Mean |       Error |    StdDev | Ratio | RatioSD |     Gen 0 |     Gen 1 | Gen 2 |    Allocated |
|---------------------- |----------------------------------------------- |-------------:|------------:|----------:|------:|--------:|----------:|----------:|------:|-------------:|
| 'System.Drawing Tiff' |                          Tiff/CCITTGroup4.tiff |   1,529.8 us |   568.26 us |  31.15 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |                          Tiff/CCITTGroup4.tiff | 117,745.8 us | 4,076.57 us | 223.45 us | 76.99 |    1.64 |         - |         - |     - |     20,872 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |    Tiff/Calliphora_grayscale_uncompressed.tiff |     196.0 us |   774.67 us |  42.46 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |    Tiff/Calliphora_grayscale_uncompressed.tiff |     176.2 us |   175.31 us |   9.61 us |  0.94 |    0.26 |         - |         - |     - |     22,688 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |     Tiff/Calliphora_rgb_deflate_predictor.tiff |   1,195.0 us |   551.90 us |  30.25 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |     Tiff/Calliphora_rgb_deflate_predictor.tiff |   1,082.5 us |   424.99 us |  23.30 us |  0.91 |    0.03 |         - |         - |     - |     23,856 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |         Tiff/Calliphora_rgb_lzw_predictor.tiff |   1,212.8 us |   663.35 us |  36.36 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |         Tiff/Calliphora_rgb_lzw_predictor.tiff |   5,099.8 us | 3,391.88 us | 185.92 us |  4.20 |    0.03 |         - |         - |     - |  2,935,984 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |              Tiff/Calliphora_rgb_packbits.tiff |     472.1 us |   209.06 us |  11.46 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |              Tiff/Calliphora_rgb_packbits.tiff |     206.3 us |   364.32 us |  19.97 us |  0.44 |    0.04 |         - |         - |     - |     22,840 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' | Tiff/Calliphora_rgb_palette_lzw_predictor.tiff |  34,671.3 us | 8,561.18 us | 469.27 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' | Tiff/Calliphora_rgb_palette_lzw_predictor.tiff |  43,881.6 us | 1,559.07 us |  85.46 us |  1.27 |    0.02 | 4000.0000 | 1000.0000 |     - | 18,032,064 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |          Tiff/Calliphora_rgb_uncompressed.tiff |     269.9 us |   114.12 us |   6.26 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |          Tiff/Calliphora_rgb_uncompressed.tiff |     170.5 us |   247.64 us |  13.57 us |  0.63 |    0.04 |         - |         - |     - |     22,736 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |     Tiff/ccitt_fax3_all_terminating_codes.tiff |     138.0 us |   881.60 us |  48.32 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |     Tiff/ccitt_fax3_all_terminating_codes.tiff |     493.3 us |   407.67 us |  22.35 us |  3.80 |    0.98 |         - |         - |     - |     19,320 B |
|                       |                                                |              |             |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |         Tiff/huffman_rle_all_makeup_codes.tiff |     118.6 us |   219.62 us |  12.04 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |         Tiff/huffman_rle_all_makeup_codes.tiff |   2,893.1 us |    46.58 us |   2.55 us | 24.56 |    2.41 |         - |         - |     - |     19,472 B |
```

PR

```
|                Method |                                      TestImage |        Mean |      Error |    StdDev | Ratio | RatioSD |     Gen 0 |     Gen 1 | Gen 2 |    Allocated |
|---------------------- |----------------------------------------------- |------------:|-----------:|----------:|------:|--------:|----------:|----------:|------:|-------------:|
| 'System.Drawing Tiff' |                          Tiff/CCITTGroup4.tiff |  1,557.2 us |   521.1 us |  28.56 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |                          Tiff/CCITTGroup4.tiff | 26,244.3 us | 1,184.6 us |  64.93 us | 16.86 |    0.35 |         - |         - |     - |     20,872 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |    Tiff/Calliphora_grayscale_uncompressed.tiff |    156.7 us |   302.2 us |  16.56 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |    Tiff/Calliphora_grayscale_uncompressed.tiff |    184.0 us |   290.1 us |  15.90 us |  1.19 |    0.22 |         - |         - |     - |     22,688 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |     Tiff/Calliphora_rgb_deflate_predictor.tiff |  1,194.3 us |   769.5 us |  42.18 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |     Tiff/Calliphora_rgb_deflate_predictor.tiff |  1,113.0 us |   989.2 us |  54.22 us |  0.93 |    0.07 |         - |         - |     - |     23,856 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |         Tiff/Calliphora_rgb_lzw_predictor.tiff |  1,269.0 us |   892.3 us |  48.91 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |         Tiff/Calliphora_rgb_lzw_predictor.tiff |  4,957.7 us | 1,098.1 us |  60.19 us |  3.91 |    0.14 |         - |         - |     - |  2,935,984 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |              Tiff/Calliphora_rgb_packbits.tiff |    488.5 us |   673.0 us |  36.89 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |              Tiff/Calliphora_rgb_packbits.tiff |    193.9 us |   303.3 us |  16.62 us |  0.40 |    0.00 |         - |         - |     - |     22,840 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' | Tiff/Calliphora_rgb_palette_lzw_predictor.tiff | 35,581.1 us | 5,135.7 us | 281.50 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' | Tiff/Calliphora_rgb_palette_lzw_predictor.tiff | 44,538.7 us | 2,855.5 us | 156.52 us |  1.25 |    0.01 | 4000.0000 | 1000.0000 |     - | 18,032,064 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |          Tiff/Calliphora_rgb_uncompressed.tiff |    288.1 us |   139.6 us |   7.65 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |          Tiff/Calliphora_rgb_uncompressed.tiff |    173.3 us |   467.9 us |  25.65 us |  0.60 |    0.07 |         - |         - |     - |     22,736 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |     Tiff/ccitt_fax3_all_terminating_codes.tiff |    119.2 us |   236.1 us |  12.94 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |     Tiff/ccitt_fax3_all_terminating_codes.tiff |    454.9 us |   580.4 us |  31.81 us |  3.83 |    0.14 |         - |         - |     - |     19,320 B |
|                       |                                                |             |            |           |       |         |           |           |       |              |
| 'System.Drawing Tiff' |         Tiff/huffman_rle_all_makeup_codes.tiff |    121.1 us |   145.5 us |   7.98 us |  1.00 |    0.00 |         - |         - |     - |        648 B |
|     'ImageSharp Tiff' |         Tiff/huffman_rle_all_makeup_codes.tiff |  1,050.5 us |   850.1 us |  46.60 us |  8.69 |    0.54 |         - |         - |     - |     19,472 B |
```

note: `Tiff/CCITTGroup4.tiff` is the image from #2132