@TonyValenti 2022-05-29
@brianpopow -
I was looking at this code and I noticed something I wanted to mention to you.

I noticed that there seem to be a number of "constant" dictionaries that are used. 

I've seen many situations where dictionaries like that are refactored into a static switch method.  This provides better performance because the C# compiler will optimize each switch statement instead of relying on the General purpose Dictionary Class.

Wanted to mention that because it may further reduce memory as well.