@TonyValenti 2022-05-29
Err, further reduce execution time.