@brianpopow 2022-05-30
> @brianpopow - I was looking at this code and I noticed something I wanted to mention to you.
> 
> I noticed that there seem to be a number of "constant" dictionaries that are used.
> 
> I've seen many situations where dictionaries like that are refactored into a static switch method. This provides better performance because the C# compiler will optimize each switch statement instead of relying on the General purpose Dictionary Class.
> 
> Wanted to mention that because it may further reduce memory as well.

So far we are down from **117,745.8 us** to **15,028.70 us**. 
I think that this is not bad, but the difference to System.Drawing is with **1,582.76 us** still quite a bit. I am currently out of idea's how to improve that further.
I tried avoiding dictionary's with d5a3de11, it's only a very small difference unfortunately.