@br3aker 2022-06-01
I'll take a closer look and play with benchmarks when this is shipped but I don't understand atm why bits are pushed bit by bit when we can have a lookup table or maybe a somehow optimized byte value generation.