@br3aker 2022-06-02
> Nevertheless I think the improvement compared to main is already quite significant.

90% speedup is simply incredible :)

I believe we can win a lot in those bit insertions though.