@TonyValenti 2022-06-09
@brianpopow Thank you so much for your work on this.

Do you have the ability to publish an updated nuget package?  If so, can you make that happen?